+++ src/internal/services/authentication/components/http_access_system/internal.go
@@ -1,0 +1,7 @@
+package http_access_system
+
+// loadHttpRoutesInRedisDb - загрузка http маршрутов в базу данных redis.
+func (acc *accessSystem) loadHttpRoutesInRedisDb() (err error) {
+
+	return
+}