+++ src/pkg/core/components/tracer/logger/config/config.go
@@ -76,7 +76,7 @@ func (conf *Config) Default() *Config {
 					TimeFormat: time.RFC3339,
 					Rotation: &FilesLogFileOptionRotation{
 						Enable:   true,
-						FileSize: "4GB",
+						FileSize: "1GB",
 					},
 				},