+++ var/www/dashboard/client/_app/immutable/nodes/5.2X5ARYQd.js
@@ -1,0 +1,35 @@
+import{s as Y,g as Z,o as tt,n as et,a as rt}from"../chunks/scheduler.VUuLSjLj.js";import{S as st,i as lt,j as q,b as A,m as o,t as nt,n as K,o as at,d as e,k as G,f as P,g as n,p as c,c as it,q as N,h as t,a as O,l,r as Q,u as R,v as U,w as W}from"../chunks/index.YcojCQiS.js";import{T as X,G as ot}from"../chunks/Text.7olNyAOr.js";function ct($){let w,m,f,a,r,h,g,_,u,d,i,E,b,B,F,k,T,L,v,V,D,x,S;return document.title=w=$[0],v=new X({props:{key:"dashboard.pages.errors.403.title",source:$[1]}}),x=new X({props:{key:"dashboard.pages.errors.403.description",source:$[1]}}),{c(){m=q(),f=A("div"),a=A("div"),r=o("svg"),h=o("g"),g=o("g"),_=o("g"),u=o("style"),d=nt(`.isometric_een{fill:rgb(var(--neutral-fg));} \r
+                        .isometric_twee{fill:rgb(var(--smoke));} \r
+                        .isometric_vier{fill:rgb(var(--neutral));} \r
+                        .isometric_twaalf{fill:#569080;} \r
+                        .isometric_dertien{fill:#225B49;} \r
+                        .st0{fill:#7BD6C4;} \r
+                        .st1{fill:#F05A28;} \r
+                        .st2{fill:#FFBB33;} \r
+                        .st3{fill:#BE1E2D;} \r
+                        .st4{fill:#F29227;} \r
+                        .st5{fill:#FF7344;} \r
+                        .st6{fill:#6B9086;} \r
+                        .st7{fill:none;} \r
+                        .st8{fill:#72C0AB;} \r
+                        .st9{fill:#F2D76C;} \r
+                        .st10{fill:#F28103;} \r
+                        .st11{fill:#225B49;} \r
+                    `),i=o("g"),E=o("path"),b=o("polygon"),B=o("path"),F=o("path"),k=o("path"),T=q(),L=A("p"),K(v.$$.fragment),V=q(),D=A("p"),K(x.$$.fragment),this.h()},l(s){at("svelte-1uo06u1",document.head).forEach(e),m=G(s),f=P(s,"DIV",{class:!0});var M=n(f);a=P(M,"DIV",{class:!0});var p=n(a);r=c(p,"svg",{class:!0,width:!0,version:!0,xmlns:!0,"xmlns:xlink":!0,viewBox:!0,"xml:space":!0,fill:!0});var z=n(r);h=c(z,"g",{"stroke-width":!0}),n(h).forEach(e),g=c(z,"g",{"stroke-linecap":!0,"stroke-linejoin":!0}),n(g).forEach(e),_=c(z,"g",{});var j=n(_);u=c(j,"style",{type:!0});var I=n(u);d=it(I,`.isometric_een{fill:rgb(var(--neutral-fg));} \r
+                        .isometric_twee{fill:rgb(var(--smoke));} \r
+                        .isometric_vier{fill:rgb(var(--neutral));} \r
+                        .isometric_twaalf{fill:#569080;} \r
+                        .isometric_dertien{fill:#225B49;} \r
+                        .st0{fill:#7BD6C4;} \r
+                        .st1{fill:#F05A28;} \r
+                        .st2{fill:#FFBB33;} \r
+                        .st3{fill:#BE1E2D;} \r
+                        .st4{fill:#F29227;} \r
+                        .st5{fill:#FF7344;} \r
+                        .st6{fill:#6B9086;} \r
+                        .st7{fill:none;} \r
+                        .st8{fill:#72C0AB;} \r
+                        .st9{fill:#F2D76C;} \r
+                        .st10{fill:#F28103;} \r
+                        .st11{fill:#225B49;} \r
+                    `),I.forEach(e),i=c(j,"g",{});var C=n(i);E=c(C,"path",{class:!0,d:!0}),n(E).forEach(e),b=c(C,"polygon",{class:!0,points:!0}),n(b).forEach(e),B=c(C,"path",{class:!0,d:!0}),n(B).forEach(e),F=c(C,"path",{class:!0,d:!0}),n(F).forEach(e),k=c(C,"path",{class:!0,d:!0}),n(k).forEach(e),C.forEach(e),j.forEach(e),z.forEach(e),T=G(p),L=P(p,"P",{class:!0});var H=n(L);N(v.$$.fragment,H),H.forEach(e),V=G(p),D=P(p,"P",{class:!0});var J=n(D);N(x.$$.fragment,J),J.forEach(e),p.forEach(e),M.forEach(e),this.h()},h(){t(h,"stroke-width","0"),t(g,"stroke-linecap","round"),t(g,"stroke-linejoin","round"),t(u,"type","text/css"),t(E,"class","isometric_vier"),t(E,"d","M21.08,14.291c0-0.124,0,11.548,0,11.548L27,21.1v-9.987L21.08,14.291z"),t(b,"class","isometric_twee"),t(b,"points","21.08,14.167 7.695,6.456 5,7.963 18.4,15.711 18.4,28 21.08,26.453 "),t(B,"class","isometric_een"),t(B,"d","M8.545,6.975l7.624-2.107L27,11.113l-5.92,3.178c-0.002-0.08,0.003-0.125,0-0.124L8.545,6.975z M5,7.963L5,7.963L5,7.963v12.301L18.4,28V15.711L5,7.963z"),t(F,"class","isometric_dertien"),t(F,"d","M6.546,10.791v8.828l9.642,5.567C13.221,22.645,7.166,18.618,6.546,10.791z"),t(k,"class","isometric_twaalf"),t(k,"d","M16.853,16.344L6.546,10.393v0.398c0.605,7.644,6.174,11.425,9.642,14.394l0.666,0.384V16.344z"),t(r,"class","monitor mx-auto svelte-1fg87d"),t(r,"width","300"),t(r,"version","1.1"),t(r,"xmlns","http://www.w3.org/2000/svg"),t(r,"xmlns:xlink","http://www.w3.org/1999/xlink"),t(r,"viewBox","0 0 32 32"),t(r,"xml:space","preserve"),t(r,"fill","#000000"),t(L,"class","text-center font-semibold text-5xl"),t(D,"class","text-center text-neutral-400 font-semibold text-xl mt-5"),t(a,"class","m-auto"),t(f,"class","fs-page")},m(s,y){O(s,m,y),O(s,f,y),l(f,a),l(a,r),l(r,h),l(r,g),l(r,_),l(_,u),l(u,d),l(_,i),l(i,E),l(i,b),l(i,B),l(i,F),l(i,k),l(a,T),l(a,L),Q(v,L,null),l(a,V),l(a,D),Q(x,D,null),S=!0},p(s,[y]){(!S||y&1)&&w!==(w=s[0])&&(document.title=w);const M={};y&2&&(M.source=s[1]),v.$set(M);const p={};y&2&&(p.source=s[1]),x.$set(p)},i(s){S||(R(v.$$.fragment,s),R(x.$$.fragment,s),S=!0)},o(s){U(v.$$.fragment,s),U(x.$$.fragment,s),S=!1},d(s){s&&(e(m),e(f)),W(v),W(x)}}}function ft($,w,m){let f,a=et,r=()=>(a(),a=rt(d,i=>m(3,f=i)),d);$.$$.on_destroy.push(()=>a());let h="";const g=function(){d&&m(0,h=ot("dashboard.pages.errors.403.title",f))},_=Z("i18nLoadDictionary");let u,d;return tt(async()=>{m(1,u=_(["dashboard.pages.errors.403"])),u.then(i=>{r(m(2,d=i)),g()})}),[h,u,d]}class pt extends st{constructor(w){super(),lt(this,w,ft,ct,Y,{})}}export{pt as component};