+++ var/www/dashboard/server/chunks/1-BX5mAPtW.js
@@ -1,0 +1,9 @@
+const index = 1;
+let component_cache;
+const component = async () => component_cache ??= (await import('./error.svelte-D4TV4xjl.js')).default;
+const imports = ["_app/immutable/nodes/1.CRpcYjGU.js","_app/immutable/chunks/scheduler.VUuLSjLj.js","_app/immutable/chunks/index.YcojCQiS.js","_app/immutable/chunks/entry.BtcyXRKn.js","_app/immutable/chunks/index.oxanXEPY.js"];
+const stylesheets = [];
+const fonts = [];
+
+export { component, fonts, imports, index, stylesheets };
+//# sourceMappingURL=1-BX5mAPtW.js.map