+++ var/www/dashboard/server/chunks/2-CR6WVYfP.js
@@ -1,0 +1,9 @@
+const index = 2;
+let component_cache;
+const component = async () => component_cache ??= (await import('./_layout.svelte-D1ZYthc_.js')).default;
+const imports = ["_app/immutable/nodes/2.BSRpFAcz.js","_app/immutable/chunks/scheduler.VUuLSjLj.js","_app/immutable/chunks/index.YcojCQiS.js","_app/immutable/chunks/index.DbE-JpqF.js","_app/immutable/chunks/index.oxanXEPY.js","_app/immutable/chunks/Button.CAzrSwEs.js"];
+const stylesheets = ["_app/immutable/assets/2.B2Qo3GWJ.css","_app/immutable/assets/Button.8CCOSIuk.css"];
+const fonts = [];
+
+export { component, fonts, imports, index, stylesheets };
+//# sourceMappingURL=2-CR6WVYfP.js.map