+++ var/www/dashboard/server/chunks/3-DO7jw3Gt.js
@@ -1,0 +1,9 @@
+const index = 3;
+let component_cache;
+const component = async () => component_cache ??= (await import('./_page.svelte-upeK2d2N.js')).default;
+const imports = ["_app/immutable/nodes/3.COaEHoVv.js","_app/immutable/chunks/scheduler.VUuLSjLj.js","_app/immutable/chunks/index.YcojCQiS.js"];
+const stylesheets = [];
+const fonts = [];
+
+export { component, fonts, imports, index, stylesheets };
+//# sourceMappingURL=3-DO7jw3Gt.js.map