+++ var/www/dashboard/server/chunks/4-C5dAoAwW.js
@@ -1,0 +1,9 @@
+const index = 4;
+let component_cache;
+const component = async () => component_cache ??= (await import('./_page.svelte-A_y5iaYC.js')).default;
+const imports = ["_app/immutable/nodes/4.B8ZQbYEy.js","_app/immutable/chunks/scheduler.VUuLSjLj.js","_app/immutable/chunks/index.YcojCQiS.js","_app/immutable/chunks/index.oxanXEPY.js","_app/immutable/chunks/index.DbE-JpqF.js","_app/immutable/chunks/index.BzIH2BPK.js","_app/immutable/chunks/Text.7olNyAOr.js","_app/immutable/chunks/Button.CAzrSwEs.js"];
+const stylesheets = ["_app/immutable/assets/4.B2EIks30.css","_app/immutable/assets/index.qJOibVgg.css","_app/immutable/assets/Button.8CCOSIuk.css"];
+const fonts = [];
+
+export { component, fonts, imports, index, stylesheets };
+//# sourceMappingURL=4-C5dAoAwW.js.map