+++ var/www/dashboard/server/chunks/_page.svelte-upeK2d2N.js
@@ -1,0 +1,8 @@
+import { c as create_ssr_component } from './ssr-DMU0SlpY.js';
+
+const Page = create_ssr_component(($$result, $$props, $$bindings, slots) => {
+  return ``;
+});
+
+export { Page as default };
+//# sourceMappingURL=_page.svelte-upeK2d2N.js.map